requirements.txt
numpy
iio==0.0.3
Pillow==9.0.0
torch
torchvision
matplotlib==3.5.3
seaborn==0.11.2
scikit-learn==1.0.2
imageio-ffmpeg==0.4.7
ruptures==1.1.5
scikit-image
pandas